requirements.txt
streamlit==1.38.0
pdfplumber==0.11.4
python-docx==1.1.2
pandas==2.2.2
numpy==2.1.1
pydantic==2.8.2
nltk==3.9.1
spacy==3.7.5
pytextrank==3.3.0
rapidfuzz==3.9.6
pytesseract==0.3.13
pillow==10.4.0
pymupdf==1.24.9
reportlab==4.2.5
fastapi==0.115.0
uvicorn==0.30.6
pypdf==5.0.1
python-magic-bin==0.4.14
cryptography==43.0.3
slowapi==0.1.9
requests==2.32.3
pdf2image==1.17.0
transformers==4.44.2
torch==2.4.1
scikit-learn==1.5.2
datasets==2.21.0
evaluate==0.4.2
onnxruntime==1.19.2
captum==0.7.0
shap==0.46.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
prometheus-client==0.20.0
scipy==1.13.1
boto3==1.35.36
clamd==1.0.2
